Revert "3271 adding calculated field replication type to experiment"
