+++ src/encoded/audit/experiment.py
@@ -89,161 +89,6 @@ def audit_experiment_replicated(value, system):
             raise AuditFailure('unreplicated experiment', detail, level='WARNING')
 
 
-@audit_checker('experiment', frame=['replicates', 'replicates.library.biosample','replicates.library.biosample.donor', 'replicates.library.biosample.donor.organism' ])
-def audit_experiment_isogeneity(value, system):
-
-    if value['status'] in ['deleted']:
-        return
-        
-    if len(value['replicates'])==0:
-        return 
-
-    biological_rep_biosample_dict = {}
-
-    for rep in value['replicates']:
-        bio_rep_num = rep['biological_replicate_number']
-        tech_rep_num = rep['technical_replicate_number']
-
-        if 'library' not in rep:
-            continue
-        lib = rep['library']
-        if 'biosample' not in lib: # checked at the level of library            
-            continue
-        biosample = lib['biosample']
-        if 'donor' not in biosample: # already checked at biosample
-            continue 
-            
-
-        if not bio_rep_num in biological_rep_biosample_dict:
-            biological_rep_biosample_dict[bio_rep_num]={}
-        biological_rep_biosample_dict[bio_rep_num][tech_rep_num]=biosample
-     
-    if len(biological_rep_biosample_dict.keys())<1:
-        return   
-    ''' 
-    To allow validation of biological replicates - first make sure technical replicates are from the same donor
-    '''
-    for tech_rep_dict_key in biological_rep_biosample_dict.keys():
-        donorsList = []
-        for tech_rep_donor_key in biological_rep_biosample_dict[tech_rep_dict_key].keys():
-            donorsList.append(biological_rep_biosample_dict[tech_rep_dict_key][tech_rep_donor_key]['donor'])
-        
-        if len(donorsList)>1:
-            initialDonor = donorsList[0]
-            for tech_donor in donorsList:
-                if initialDonor['accession']!=tech_donor['accession']:  
-                    #######################################################################################
-                    # talk with Sricket about the return value in case of anisogenic technical replicates #
-                    #######################################################################################                  
-                    raise technical_failure(value)
-    
-    '''
-    Now we can proceed to biological replicates validation
-    '''
-    bio_reps_biosamples = []
-    for tech_rep_dict_key in biological_rep_biosample_dict.keys():
-        for tech_rep_donor_key in biological_rep_biosample_dict[tech_rep_dict_key].keys():
-            bio_rep_biosample = biological_rep_biosample_dict[tech_rep_dict_key][tech_rep_donor_key]
-            break
-        bio_reps_biosamples.append(bio_rep_biosample) #adding donor
-
-    '''
-    In the bio_reps we have a list of representative bilogical donors
-    Now we can check different things for humans or model organisms
-    '''    
-    initialBiosample = bio_reps_biosamples[0]
-    initialOrganism = initialBiosample['donor']['organism']
-    initialAccession = initialBiosample['donor']['accession']
-    
-    humanFlag = False
-    if initialOrganism['scientific_name']=='Homo sapiens': #dealing with humans for now
-        humanFlag = True
-    for biosample_entry in bio_reps_biosamples:
-        if biosample_entry['donor']['accession'] != initialAccession: # we have mismatched biological replicates
-            if 'sex' in biosample_entry and 'age' in biosample_entry: # now we have to figure out if there is a match in age/sex
-                ageMatchBoolean = isMatchedAge(biosample_entry, initialBiosample)
-                sexMatchedBoolean = isMatchedSex(biosample_entry, initialBiosample, humanFlag)
-
-                if ageMatchBoolean == True and sexMatchedBoolean == True:
-                    return biological_matched_sex_age(value, initialBiosample, biosample_entry)
-                if ageMatchBoolean == False and sexMatchedBoolean == True:
-                    return biological_mismatched_age(value, initialBiosample, biosample_entry)
-                if ageMatchBoolean == True and sexMatchedBoolean == False:
-                    return biological_mismatched_sex(value, initialBiosample, biosample_entry)
-                if ageMatchBoolean == False and sexMatchedBoolean == False:
-                    return biological_mismatched_sex_age(value, initialBiosample, biosample_entry)
-            else:
-                return biological_mismatched_sex_age(value, initialBiosample, biosample_entry)
-                    
-def isMatchedAge(biosample_one, biosample_two):
-    age_1 = biosample_one['age']
-    age_2 = biosample_two['age']    
-
-    if age_1 == 'unknown' or age_2 == 'uknown':
-        return False
-
-    age_1_units = biosample_one['age_units']
-    age_2_units = biosample_one['age_units']
-
-    if age_1_units == age_2_units and age_1 == age_2:
-        return True
-
-    return False
-
-def isMatchedSex(biosample_one, biosample_two, isHuman):
-    sex_1 = biosample_one['sex']
-    sex_2 = biosample_two['sex']
-
-    #cover unknown case
-    if sex_1 == 'unknown' or sex_2 == 'unknown':
-        return False
-    
-    # equal sexes [(]male,female, herms, but not unknown (see above) and not mixed]
-    if sex_1 == sex_2 and sex_1 != 'mixed':
-        return True
-
-    # equal mixed sex for model organisms
-    if isHuman==False:
-        if sex_1 == sex_2 and sex_1 == 'mixed':
-            return True 
-
-    return False
-
-def technical_failure(value):
-    detail = 'Experiment {} has anisogenic technical replicates'.format(value['@id'])
-    return AuditFailure('anisogenic technical replicates', detail, level='DCC_ACTION')
-
-def biological_matched_sex_age(value, biosample_one, biosample_two):
-    detail = 'In experiment {} the biological replicates for biosamples {} and {} are anisogenic, sex and age properties are matched'.format(
-        value['@id'],
-        biosample_one['@id'],
-        biosample_two['@id'])
-    yield AuditFailure('anisogenic biological replicates, matched sex and age', detail, level='DCC_ACTION')
-
-def biological_mismatched_sex_age(value, biosample_one, biosample_two):
-    detail = 'In experiment {} the biological replicates for biosamples {} and {} are anisogenic, sex and age properties are not matched'.format(
-        value['@id'],
-        biosample_one['@id'],
-        biosample_two['@id'])
-    yield AuditFailure('anisogenic biological replicates, mismatched sex and age', detail, level='DCC_ACTION')
-
-def biological_mismatched_sex(value, biosample_one, biosample_two):
-    detail = 'In experiment {} the biological replicates for biosamples {} and {} are anisogenic, sex property is not matched'.format(
-        value['@id'],
-        biosample_one['@id'],
-        biosample_two['@id'])
-    yield AuditFailure('anisogenic biological replicates, matched age', detail, level='DCC_ACTION')
-
-def biological_mismatched_age(value, biosample_one, biosample_two):
-    detail = 'In experiment {} the biological replicates for biosamples {} and {} are anisogenic, age property is not matched'.format(
-        value['@id'],
-        biosample_one['@id'],
-        biosample_two['@id'])
-    yield AuditFailure('anisogenic biological replicates, matched sex', detail, level='DCC_ACTION')
-
-
-
-
 @audit_checker('experiment', frame=['replicates', 'replicates.library'])
 def audit_experiment_technical_replicates_same_library(value, system):
     if value['status'] in ['deleted', 'replaced', 'revoked']:
