+++ src/encoded/audit/library.py
@@ -14,19 +14,6 @@ moleculeDict = {
     }
 
 
-@audit_checker('library', frame=['object'])
-def audit_library_biosample(value, system):
-    '''
-    The library should be linked to biosample 
-    '''
-    if value['status'] in ['deleted']:
-        return
-    if 'biosample' not in value:
-        detail = 'Library {} has no biosample'.format(
-            value['@id'])
-        raise AuditFailure('missing biosample', detail, level='ERROR')
-
-
 @audit_checker('library', frame='object')
 def audit_library_nucleic_acid(value, system):
     '''
@@ -71,8 +58,6 @@ def audit_library_documents(value, system):
                 value['@id']
                 )
             raise AuditFailure('missing documents', detail, level='WARNING')
-
-
 
 
 @audit_checker('library', frame='object')
